+++ src/cspark/sdk/resources/__init__.py
@@ -1,4 +1,6 @@
 from ._base import *
 from ._batches import *
+from ._history import *
+from ._impex import *
 from ._oauth2 import *
 from ._services import *