+++ src/cspark/sdk/resources/_base.py
@@ -64,10 +64,11 @@ class ApiResource:
         method: str = 'GET',
         headers: Mapping[str, str] = {},
         params: Optional[Mapping[str, str]] = None,
-        body=None,
+        body: Optional[Any] = None,
+        content: Optional[bytes] = None,
         form=None,
         files=None,
-    ):
+    ) -> HttpResponse:
         url = url.value if isinstance(url, Uri) else url
         request = self._client.build_request(
             method,
@@ -76,6 +77,7 @@ class ApiResource:
             headers={**headers, **self.default_headers},
             data=form,
             json=body,
+            content=content,
             files=files,
             timeout=self.config.timeout / 1000,
         )
@@ -83,9 +85,9 @@ class ApiResource:
         self.logger.debug(f'{method} {url}')
         return self.__fetch(request)
 
-    def __fetch(self, request: Request, retries: int = 0) -> 'HttpResponse':
+    def __fetch(self, request: Request, retries: int = 0) -> HttpResponse:
         request.headers.update(self.config.auth.as_header)
-        response = self._client.send(request)
+        response = self._client.send(request)  # FIXME: handling errors (httpx.RequestError | httpx.HTTPStatusError)
         status = response.status_code
 
         if status >= 400:
@@ -113,8 +115,8 @@ class ApiResource:
                         },
                         'response': {
                             'headers': response.headers,
-                            'body': response.content,  # FIXME: cast to dict if possible
-                            'raw': response.text,
+                            'body': response.text,
+                            'raw': response.content,
                         },
                     },
                 },
@@ -129,6 +131,52 @@ class ApiResource:
                 http_response['data'] = response.json()
             except Exception:
                 http_response['data'] = response.text
+        return HttpResponse(**http_response)
+
+
+def download(
+    url: str,
+    *,
+    method: str = 'GET',
+    headers: Mapping[str, str] = {},
+    params: Optional[Mapping[str, str]] = None,
+    body=None,
+    form=None,
+    timeout: Optional[float] = 60,
+):
+    with Client() as client:
+        request = client.build_request(
+            method,
+            url,
+            params=params,
+            headers=headers,
+            data=form,
+            json=body,
+            timeout=timeout,
+        )
+        response = client.send(request)
+        status = response.status_code
+        if status >= 400:
+            raise SparkError.api(
+                response.status_code,
+                {
+                    'message': f'failed to download content from <{request.url}>',
+                    'cause': {
+                        'request': {
+                            'url': str(request.url),
+                            'method': request.method,
+                            'headers': request.headers,
+                            'body': request.content,
+                        },
+                        'response': {
+                            'headers': response.headers,
+                            'body': response.text,
+                            'raw': response.content,
+                        },
+                    },
+                },
+            )
+        http_response = {'status': status, 'data': None, 'buffer': response.content, 'headers': response.headers}
         return HttpResponse(**http_response)